Move ColorPicker components into a separate folder
