+++ src/common/components/ColorPicker/ColorPickerDialog.jsx
